SMOOTH MOVEMENT!

+++ src/game/GameObject.h
@@ -4,7 +4,7 @@
 class GameObject{
 	public:
 		virtual void display()=0;
-		virtual void gameUpdate(float time)=0;
+		virtual void gameUpdate(Uint32 time)=0;
 		
 		virtual ~GameObject(){};
 };
